requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Admin==1.6.1
Flask-WTF==1.2.1
WTForms==3.1.1
Flask-Bcrypt==1.0.1
Flask-Uploads==0.2.1
Flask-Talisman==1.1.0
python-dotenv==1.0.0
Pillow==10.1.0
email-validator==2.1.0
pymysql==1.1.2
cryptography==46.0.3
